updated triangle class
