+++ lib/shapes.js
@@ -4,16 +4,18 @@ class Shape {
         this.textColor = textColor;
     }
     render()    {
-        if ((this.text).length > 3) {
-            throw new Error("Text Must not be more than 3 letters.");
-        }
+        throw new Error("Text Must not be more than 3 letters.");
     }
     renderSVGTag()  {
         return `<svg xmlns="http://www.w3.org/2000/svg" version="1.1" width="300" height="300">`
     }
     renderTitleTag()    {
-        return `<text x="50%" y="50%" text-anchor ="middle" font-size="100px" fill = ${this.textColor}>${this.text}</text>
-        </svg>`
+        if ((this.text).length > 3) {
+            this.render();
+        } else {
+            return `<text x="50%" y="50%" text-anchor ="middle" font-size="90px" fill = "${this.textColor}">${this.text}</text>
+            </svg>`
+        }
     }
